[PIECE-61] State, Intnet, SideEffect contract 패키지로 이동 및 intent Channel 추가
