+++ feature/matching/src/main/java/com/puzzle/matching/contract/MatchingSideEffect.kt
@@ -1,0 +1,3 @@
+package com.puzzle.matching.contract
+
+sealed class MatchingSideEffect
