+++ feature/matching/src/main/java/com/puzzle/matching/contract/MatchingState.kt
@@ -1,0 +1,7 @@
+package com.puzzle.matching.contract
+
+import com.airbnb.mvrx.MavericksState
+
+data class MatchingState(
+    val isLoading: Boolean = false,
+) : MavericksState
