+++ feature/matching/src/main/java/com/puzzle/matching/detail/MatchingDetailRoute.kt
@@ -10,6 +10,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.sp
 import com.airbnb.mvrx.compose.collectAsState
 import com.airbnb.mvrx.compose.mavericksViewModel
+import com.puzzle.matching.detail.contract.MatchingDetailState
 
 @Composable
 fun MatchingDetailRoute(
