+++ feature/matching/src/main/java/com/puzzle/matching/detail/MatchingDetailViewModel.kt
@@ -1,30 +1,42 @@
 package com.puzzle.matching.detail
 
-import com.airbnb.mvrx.MavericksState
 import com.airbnb.mvrx.MavericksViewModel
 import com.airbnb.mvrx.MavericksViewModelFactory
 import com.airbnb.mvrx.hilt.AssistedViewModelFactory
 import com.airbnb.mvrx.hilt.hiltMavericksViewModelFactory
-import com.puzzle.navigation.NavigationHelper
+import com.puzzle.matching.detail.contract.MatchingDetailIntent
+import com.puzzle.matching.detail.contract.MatchingDetailSideEffect
+import com.puzzle.matching.detail.contract.MatchingDetailState
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedFactory
 import dagger.assisted.AssistedInject
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.channels.Channel.Factory.BUFFERED
+import kotlinx.coroutines.flow.launchIn
+import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.flow.receiveAsFlow
-
-data class MatchingDetailState(
-    val isLoading: Boolean = false,
-) : MavericksState
+import kotlinx.coroutines.launch
 
 class MatchingDetailViewModel @AssistedInject constructor(
     @Assisted initialState: MatchingDetailState,
 ) : MavericksViewModel<MatchingDetailState>(initialState) {
 
+    private val intents = Channel<MatchingDetailIntent>(BUFFERED)
+
     private val _sideEffect = Channel<MatchingDetailSideEffect>(BUFFERED)
     val sideEffect = _sideEffect.receiveAsFlow()
 
-    internal fun processIntent(intent: MatchingDetailIntent) {
+    init {
+        intents.receiveAsFlow()
+            .onEach(::processIntent)
+            .launchIn(viewModelScope)
+    }
+
+    internal fun onIntent(intent: MatchingDetailIntent) = viewModelScope.launch {
+        intents.send(intent)
+    }
+
+    private fun processIntent(intent: MatchingDetailIntent) {
         when (intent) {
             else -> Unit
         }
@@ -45,6 +57,3 @@ class MatchingDetailViewModel @AssistedInject constructor(
         MavericksViewModelFactory<MatchingDetailViewModel, MatchingDetailState> by hiltMavericksViewModelFactory()
 }
 
-sealed class MatchingDetailIntent
-
-sealed class MatchingDetailSideEffect
