+++ feature/matching/src/main/java/com/puzzle/matching/detail/contract/MatchingDetailIntent.kt
@@ -1,0 +1,3 @@
+package com.puzzle.matching.detail.contract
+
+sealed class MatchingDetailIntent
