+++ feature/matching/src/main/java/com/puzzle/matching/detail/contract/MatchingDetailSideEffect.kt
@@ -1,0 +1,3 @@
+package com.puzzle.matching.detail.contract
+
+sealed class MatchingDetailSideEffect
