+++ feature/matching/src/main/java/com/puzzle/matching/detail/contract/MatchingDetailState.kt
@@ -1,0 +1,7 @@
+package com.puzzle.matching.detail.contract
+
+import com.airbnb.mvrx.MavericksState
+
+data class MatchingDetailState(
+    val isLoading: Boolean = false,
+) : MavericksState
